--- src/main/java/com/alom/annotations/ValidFileExtension.java
@@ -1,22 +1,0 @@
-package com.alom.annotations;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import com.alom.annotations.impl.ExcelFileExtensionValidator;
-
-@Constraint(validatedBy = ExcelFileExtensionValidator.class)
-@Target({ ElementType.PARAMETER, ElementType.FIELD })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidFileExtension {
-    String message() default "Invalid file extension";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-    
-    // Define allowed extensions
-    String[] acceptedExtensions();
-}
