--- src/main/java/com/alom/dao/repositories/EventMediaRepository.java
@@ -1,11 +1,0 @@
-package com.alom.dao.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.alom.dao.entities.EventMediaEntity;
-
-@Repository
-public interface EventMediaRepository extends JpaRepository<EventMediaEntity, Long> {
-
-}
