+++ src/main/java/com/alom/events/annotations/ValidContactNumber.java
@@ -1,0 +1,21 @@
+package com.alom.events.annotations;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import com.alom.events.annotations.impl.ContactNumberValidator;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+@Constraint(validatedBy = ContactNumberValidator.class)
+@Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE, ElementType.PARAMETER })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ValidContactNumber {
+    String message() default "Invalid contact number";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
