+++ src/main/java/com/alom/events/annotations/ValidFileExtension.java
@@ -1,0 +1,23 @@
+package com.alom.events.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import com.alom.events.annotations.impl.ExcelFileExtensionValidator;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+@Constraint(validatedBy = ExcelFileExtensionValidator.class)
+@Target({ ElementType.PARAMETER, ElementType.FIELD })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ValidFileExtension {
+    String message() default "Invalid file extension";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+    
+    // Define allowed extensions
+    String[] acceptedExtensions();
+}
