+++ src/main/java/com/alom/events/annotations/impl/ContactNumberValidator.java
@@ -1,0 +1,24 @@
+package com.alom.events.annotations.impl;
+
+import com.alom.constant.ApiResponseMessage;
+import com.alom.events.annotations.ValidContactNumber;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
+public class ContactNumberValidator implements ConstraintValidator<ValidContactNumber, String> {
+
+    @Override
+    public void initialize(ValidContactNumber constraintAnnotation) {
+        // Initialization logic, if necessary
+    }
+
+    @Override
+    public boolean isValid(String contactNumber, ConstraintValidatorContext context) {
+        // Check if the contact number is null or empty
+        if (contactNumber == null || contactNumber.isEmpty()) {
+            return false; // Or true, depending on whether you want to allow null values
+        }
+        return contactNumber.matches(ApiResponseMessage.CONTACT_NUMBER_PATTERN);
+    }
+}
