+++ src/main/java/com/alom/events/mapper/EventMapperService.java
@@ -1,0 +1,12 @@
+package com.alom.events.mapper;
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface EventMapperService {
+
+//	EventMasterModel mappedToEventMasterDto(EventMasterEntity eventMasterEntity);
+//	List<EventMasterModel> mappedToEventMasterDto(List<EventMasterEntity> eventMasterEntity);
+//	
+//	EventMasterEntity mappedToEventMasterEntity(EventMasterModel eventMasterDto);
+}
