+++ src/main/java/com/alom/events/service/RedisService.java
@@ -1,0 +1,6 @@
+package com.alom.events.service;
+
+public interface RedisService {
+	public void addHashObject(String hashKey, String filedKey, Object data);
+	public Object retrieveHashObject(String hashKey, String filedKey);
+}
