--- src/main/java/com/alom/mapper/EventMapperService.java
@@ -1,12 +1,0 @@
-package com.alom.mapper;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface EventMapperService {
-
-//	EventMasterModel mappedToEventMasterDto(EventMasterEntity eventMasterEntity);
-//	List<EventMasterModel> mappedToEventMasterDto(List<EventMasterEntity> eventMasterEntity);
-//	
-//	EventMasterEntity mappedToEventMasterEntity(EventMasterModel eventMasterDto);
-}
