--- src/main/java/com/alom/service/EventService.java
@@ -1,27 +1,0 @@
-package com.alom.service;
-
-import java.util.Date;
-
-import org.springframework.data.domain.Pageable;
-import org.springframework.web.multipart.MultipartFile;
-
-import com.alom.model.EventMasterModel;
-import com.alom.model.PaginationResponse;
-import com.alom.payload.GenericResponse;
-
-public interface EventService {
-
-	public PaginationResponse<EventMasterModel> getAllEvents( int page, int size);
-
-	public PaginationResponse<EventMasterModel> getEventByName(String eventName, int page, int size);
-
-	public PaginationResponse<EventMasterModel> getEventBetween(Date fromEventDate, Date uptoEventDate, int page, int size);
-
-	public GenericResponse addOrUpdateEvent(MultipartFile mediaFile, MultipartFile excelFile, String jsonData);
-
-	public GenericResponse removeEventByName(String eventName);
-
-	public long getTotalEventCount();
-	
-
-}
