--- src/main/java/com/alom/service/RedisService.java
@@ -1,6 +1,0 @@
-package com.alom.service;
-
-public interface RedisService {
-	public void addHashObject(String hashKey, String filedKey, Object data);
-	public Object retrieveHashObject(String hashKey, String filedKey);
-}
